requirements.txt
Flask
flask-cors
azure-cosmos
openai
python-dotenv